Allow nova to guess device if not passed to attach
partial fix for bug 1004328

Only the xen hypervisor actually respects the device name that
is passed in attach_volume. For other hypervisors it makes much
more sense to automatically generate a unique name.

This patch generates a non-conflicting device name if one is not
passed to attach_volume. A corresponding change to novaclient
and horizon will greatly improve the user experience of attaching
a volume.

It moves some common code out of metadata/base so that it can
be used to get a list of block devices. The code was functionally
tested as well and block device name generation works properly.

The next step is to allow the user to specify an identifier that
can be used for the uuid so they can match the device in the guest
more easily.

Change-Id: I0b9454fc50a5c93b4aea38545dcee98f68d7e511