Always create the instance db record locally
A request for multiple instances was sent to the scheduler, where
it was written to the database. It appears that this was one so
that more advanced schedulers could handle the request as one
batch, but the result is the scheduler is sometimes slow enough
that the call will timeout.

Instead this converts to always creating the instance locally
and sending each request in its own cast. This should allow
large numbers of instances to be created without timouts.

Fixes bug 1036911

Change-Id: I63bbc98c285faebec53f8e62857c01548807db68