Makes sure instance deletion ok with deleted data
Commit 5ad1dea4 added changed the network deallocation code to
work with deleted instances. This was done by setting the context
to read deleted records. Unfortunately this was done a little too
broadly, leading to a new bug where a deleted floating_ip will
cause an instance to not be able to be deleted.

This fixes the issue by limiting the use of read_deleted context
to only the places it is trying to access the instance record. It
adds a test  to verify that the code works with a duplicate
deleted floating_ip and updates the existing test for a deleted
instance to exercise the entire code path.

Fixes bug 1038266
Fixes APPLIANCE-2676

Change-Id: I1aef94369e5bcf951e78e89b1eded5305cf36b53
(cherry-picked from commit 1f98e28a80077760394201f79de04a0924b9ad3f)