Override floating interface on callee side
We allow individual nodes to override the floating interface by setting
public_interface=XXX on the node. Unfortunately in multi_host an
associate or disassociate call can be forwarded from one host to
another. Currently the override happens on the caller side instead of
the callee side. This fixes the issue by moving the override.

Fixes bug 1096987

Change-Id: I37fbbf9e0ee0ffbe8c8754649301371b98da9f4f