Optimize rpc handling for allocate and deallocate
If we are in multi_host mode we can save an extra rpc call by
sending the message to the instance host the first time. This
will prevent nova-network from having to resend the message
to actually allocate the fixed ips.

Part of blueprint optimize-nova-network

DocImpact: before multi_host only affected the network create
commands, but now it can be used at run-time to optimize the
path for allocating and deallocating fixed_ips, so it is important
to set it if you are using multi_host networks.

Change-Id: Ide48e0adb2e1051e31965c9a8422b27c1a9c40b3