Fix handling of source_groups with no-db-compute.
The moving of data around in no-db-compute broke source groups. This
is a combination of a few errors. First, the instance_type data wasn't
being retrieved from system_metadata. Second, the instance_type data
was too deep in the heirarchy and being ignored. Finally, source
groups require a a nw info call which causes a db access by
nova-compute.

WIP

Change-Id: Iccb6e5d336c0b2c8ba41c41ab2e046fc9617835a